+++ geotrek/trekking/models.py
@@ -247,7 +247,10 @@ class Trek(StructureRelated, PicturesMixin, PublishableMixin, MapEntityMixin, To
 
 Path.add_property('treks', Trek.path_treks, _(u"Treks"))
 Topology.add_property('treks', Trek.topology_treks, _(u"Treks"))
-Topology.add_property('published_treks', lambda self: intersecting(Trek, self).filter(published=True), _(u"Published treks"))
+if settings.HIDE_PUBLISHED_TREKS_IN_TOPOLOGIES:
+    Topology.add_property('published_treks', lambda self: [], _(u"Published treks"))
+else:
+    Topology.add_property('published_treks', lambda self: intersecting(Trek, self).filter(published=True), _(u"Published treks"))
 Intervention.add_property('treks', lambda self: self.topology.treks if self.topology else [], _(u"Treks"))
 Project.add_property('treks', lambda self: self.edges_by_attr('treks'), _(u"Treks"))
 tourism_models.TouristicContent.add_property('treks', lambda self: intersecting(Trek, self), _(u"Treks"))